steve-translator-server/requirements.txt
Flask==1.1.2
numpy
opencv-python
Pillow